requirements.txt
beautifulsoup4==4.12.2
plyer==2.1.0
PyYAML==6.0.1
requests==2.25.1
